backend/requirements.txt
numpy
tensorflow
python-dotenv
flask
flask-cors
